Adapted test on singleton for UrlInfoBase cache (UrlInfoCollection)
Singleton add method also enures, instance is only added once.
For easy unmocked test, this will also work with ncrunch with no need
for a cleanup.
A better way could be to use the RegisteredSingleton with
IUrlInfoCollection and moch the collection for the tests, but current
impl. works like a charm ans the collection could be public for the
whole application. Addin static aliases to the UrlInfCollection would
bring more coding convinience. This could also be achieved with
extension methods on IUrlInfoCollection but is not as strict as the
concrete type. Currently we do NOT need more than the concrete type.